add student, teacher, subj controllers

+++ src/main/java/com/schedule/proj/repository/SubjectRepository.java
@@ -2,8 +2,11 @@ package com.schedule.proj.repository;
 
 import com.schedule.proj.model.Subject;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.util.Set;
+
 @Repository
-public interface SubjectRepository extends JpaRepository<Subject,Integer> {
+public interface SubjectRepository extends JpaRepository<Subject, Long> {
 }
